+++ packages/lesswrong/components/ea-forum/digestAd/StickyDigestAd.tsx
@@ -1,0 +1,198 @@
+import React from 'react';
+import { Components, registerComponent } from '../../../lib/vulcan-lib';
+import { AnalyticsContext } from '../../../lib/analyticsEvents';
+import { useCurrentUser } from '../../common/withUser';
+import { eaForumDigestSubscribeURL } from '../../recentDiscussion/RecentDiscussionSubscribeReminder';
+import { Link } from '../../../lib/reactRouterWrapper';
+import classNames from 'classnames';
+import { useDigestAd } from './useDigestAd';
+import { DIGEST_AD_BODY_TEXT, DIGEST_AD_HEADLINE_TEXT } from './SidebarDigestAd';
+
+const styles = (theme: ThemeType) => ({
+  root: {
+    animation: 'digest-fade-in 1s ease',
+    position: 'fixed',
+    bottom: 28,
+    left: '50%',
+    transform: 'translateX(-50%)',
+    width: 1100,
+    maxWidth: '85%',
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'space-between',
+    columnGap: 20,
+    backgroundColor: theme.palette.panelBackground.digestAdBanner,
+    color: theme.palette.grey[0],
+    fontSize: 13,
+    fontWeight: 450,
+    fontFamily: theme.typography.fontFamily,
+    padding: '16px 20px',
+    borderRadius: theme.borderRadius.default,
+    zIndex: theme.zIndexes.intercomButton,
+    // Once the contents start to wrap, we instead force a smaller width,
+    // so that the form doesn't get too wide
+    '@media (max-width: 880px)': {
+      display: 'block',
+      width: 500,
+      maxWidth: '90%',
+    },
+    // For now, we are not displaying this ad on small screens,
+    // since it takes up a larger percent of the screen
+    '@media (max-width: 435px)': {
+      display: 'none'
+    }
+  },
+  textCol: {
+    flexGrow: 1
+  },
+  heading: {
+    fontWeight: 600,
+    fontSize: 19,
+    margin: '0 0 4px',
+    [theme.breakpoints.down('sm')]: {
+      fontSize: 18,
+    }
+  },
+  body: {
+    fontSize: 14,
+    lineHeight: '20px',
+    fontWeight: 500,
+    color: theme.palette.grey[400],
+    textWrap: 'pretty',
+    marginBottom: 0,
+    [theme.breakpoints.down('sm')]: {
+      fontSize: 13,
+      lineHeight: '19px',
+    },
+    '@media (max-width: 880px)': {
+      marginBottom: 12,
+    },
+  },
+  form: {
+    flexGrow: 1,
+    display: 'flex',
+    columnGap: 4,
+    rowGap: '12px'
+  },
+  formInput: {
+    minWidth: 0,
+    flex: '1 1 0',
+    background: theme.palette.background.digestAdBannerInput,
+    padding: 11,
+    borderRadius: theme.borderRadius.default,
+    border: theme.palette.border.normal,
+    '&:hover': {
+      border: theme.palette.border.slightlyIntense2,
+    },
+    '&:focus': {
+      border: `1px solid ${theme.palette.primary.main}`,
+    },
+    '&::placeholder': {
+      color: theme.palette.grey[600],
+    }
+  },
+  formBtn: {
+    flex: 'none',
+  },
+  noThanksBtn: {
+    background: theme.palette.buttons.digestAdBannerNoThanks.background,
+    color: theme.palette.grey[0],
+    '&:hover': {
+      background: theme.palette.buttons.digestAdBannerNoThanks.hoverBackground,
+    },
+  },
+  success: {
+    display: 'flex',
+    alignItems: 'center',
+    columnGap: 10,
+    fontSize: 13,
+    lineHeight: '19px',
+    color: theme.palette.grey[400],
+    textWrap: 'pretty',
+  },
+  successCheckIcon: {
+    color: theme.palette.icon.greenCheckmark
+  },
+  successLink: {
+    color: theme.palette.primary.main
+  },
+});
+
+/**
+ * This is the Forum Digest ad that appears fixed to the bottom of the screen on the EA Forum post page.
+ */
+const StickyDigestAd = ({className, classes}: {
+  className?: string,
+  classes: ClassesType<typeof styles>,
+}) => {
+  const currentUser = useCurrentUser()
+  const { showDigestAd, emailRef, showForm, loading, subscribeClicked, handleClose, handleUserSubscribe } = useDigestAd()
+  
+  if (!showDigestAd) return null
+  
+  const { ForumIcon, EAButton } = Components
+  
+  const buttonProps = loading ? {disabled: true} : {}
+  const noThanksBtn = (
+    <EAButton
+      onClick={handleClose}
+      className={classNames(classes.formBtn, classes.noThanksBtn)}
+      {...buttonProps}
+    >
+      No thanks
+    </EAButton>
+  )
+  
+  // Show the form to submit to Mailchimp directly,
+  // or display the logged in user's email address and the "Sign up" button
+  let formNode = (showForm || !currentUser) ? (
+    <form action={eaForumDigestSubscribeURL} method="post" className={classes.form}>
+      <input ref={emailRef} name="EMAIL" placeholder="Email address" className={classes.formInput} required={true} />
+      <EAButton type="submit" onClick={handleUserSubscribe} className={classes.formBtn} {...buttonProps}>
+        Sign up
+      </EAButton>
+      {noThanksBtn}
+    </form>
+  ) : (
+    <div className={classes.form}>
+      <input value={currentUser.email} className={classes.formInput} disabled={true} required={true} />
+      <EAButton onClick={handleUserSubscribe} className={classes.formBtn} {...buttonProps}>
+        Sign up
+      </EAButton>
+      {noThanksBtn}
+    </div>
+  )
+  
+  // If a logged in user with an email address subscribes, show the success message.
+  if (!showForm && subscribeClicked) {
+    formNode = <div className={classes.success}>
+      <ForumIcon icon="CheckCircle" className={classes.successCheckIcon} />
+      <div>
+        Thanks for signing up! You can edit your subscription via
+        your <Link to={'/account?highlightField=subscribedToDigest'} className={classes.successLink}>
+          account settings
+        </Link>.
+      </div>
+    </div>
+  }
+  
+  return <AnalyticsContext pageSubSectionContext="digestAd">
+    <div className={classNames(classes.root, className)}>
+      <div className={classes.textCol}>
+        <h2 className={classes.heading}>{DIGEST_AD_HEADLINE_TEXT}</h2>
+        <div className={classNames(classes.body)}>
+          {DIGEST_AD_BODY_TEXT}
+        </div>
+      </div>
+      {formNode}
+    </div>
+  </AnalyticsContext>
+}
+
+const StickyDigestAdComponent = registerComponent("StickyDigestAd", StickyDigestAd, {styles, stylePriority: -1});
+
+declare global {
+  interface ComponentTypes {
+    StickyDigestAd: typeof StickyDigestAdComponent
+  }
+}
